Add license files and update social media image
Two license files have been added to the repository for software usage. The social media image in the layout file has been updated to use a new file. The README file was revised to make it more personal and clarify the purposes of the repository. Unnecessary parts, like getting started instructions and learning resources, are removed. The C-sharp logo svg file was deleted.