Update CSS and add 'Contributed Projects' section in Home module
Changes include adjusting the grid template columns in the CSS file and changing font size of certain elements. An important addition is a new section listing the projects the user has contributed to, which has been added in the Home module. The SocialButton ordering in `page.tsx` has also been revised.